fix: code refactored, get-data.js module added, modules dependencies decreased
